refactor the api file into an api folder
