+++ pages/browse/index.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from 'react'
 import { useRouter } from 'next/router'
 import { Item, SearchBar } from 'webstore-component-library'
 import { configure_services } from '../../utils'
-import { getAllWares } from '../../utils/api'
+import { getAllWares } from '../../utils/api/services'
 
 const Browse = () => {
   const router = useRouter()
