+++ utils/api/services.js
@@ -1,15 +1,5 @@
-import axios from 'axios'
 import useSWR from 'swr'
-
-const baseURL = `https://${process.env.NEXT_PUBLIC_DOMAIN_NAME}.scientist.com/api/${process.env.NEXT_PUBLIC_SCIENTIST_API_VERSION}`
-const defaultHeaders = { Authorization: `Bearer ${process.env.NEXT_PUBLIC_TOKEN}` }
-
-export const a = axios.create({ baseURL, headers: defaultHeaders })
-
-const fetcher = (...args) => {
-  return a.get(...args)
-    .then(res => res.data)
-}
+import { fetcher } from '.'
 
 export const getAllWares = () => {
   const { data, error } = useSWR(`/providers/${process.env.NEXT_PUBLIC_PROVIDER_ID}/wares.json`, fetcher)
